# courseGenie_backEnd/docs/apidocs/ui/openapi.json
{
  "openapi": "3.0.3",
  "info" : {
    "description" : "<h1>course-genie-backend<\/h1><p>Demo project for Spring Boot<\/p>",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "course-genie-backend"
  },
  "servers" : [
    {
      "url" : "/"
    }
  ],
...
  "components" : {
    "schemas" : {
...
      "SyllabusDTO" : {
"type" : "object",
"title" : "SyllabusDTO",
"required" : ["sectionId", "syllabusId"],
"properties" : {
  "content" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "sectionId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "syllabusId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  }
},
"example" : {
  "content" : "...",
  "sectionId" : 12345,
  "syllabusId" : 12345
},
"description" : ""
      }
...
      "xml_ns0_syllabusDTO" : {
"type" : "object",
"title" : "syllabusDTO",
"required" : ["sectionId", "syllabusId"],
"allOf" : [
  {
"properties" : {
  "content" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "sectionId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "syllabusId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  }
}
  }
],
"description" : ""
      }
    }
  },
  "paths": {
...
    "\/api/syllabuses" : {
      "get" : {
        "tags" : [ "SyllabusController" ],
        "operationId" : "getAllSyllabuses",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "[ {\n  \"content\" : \"...\",\n  \"sectionId\" : 12345,\n  \"syllabusId\" : 12345\n} ]"
                  }
                },
                "schema" : {
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/SyllabusDTO"
}
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<list>\n  <s-----\u003E\n    <syllabusId>...<\/syllabusId>\n    <content>...<\/content>\n    <sectionId>...<\/sectionId>\n  <\/s-----\u003E\n<\/list>\n"
                  }
                },
                "schema" : {
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/xml_ns0_syllabusDTO"
}
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
        ]
      },
      "post" : {
        "tags" : [ "SyllabusController" ],
        "operationId" : "createSyllabus",
        "requestBody" : {
          "required" : false,
          "content" : {
            "application/json" : {
              "examples" : {
                "example" : {
                  "value" : "{\n  \"content\" : \"...\",\n  \"sectionId\" : 12345,\n  \"syllabusId\" : 12345\n}"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/SyllabusDTO",
"description" : ""
              }
            }, 
            "application/xml" : {
              "examples" : {
                "example" : {
                  "value" : "<s-----\u003E\n  <syllabusId>...<\/syllabusId>\n  <content>...<\/content>\n  <sectionId>...<\/sectionId>\n<\/s-----\u003E\n"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/xml_ns0_syllabusDTO",
"description" : ""
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "{\n  \"content\" : \"...\",\n  \"sectionId\" : 12345,\n  \"syllabusId\" : 12345\n}"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/SyllabusDTO",
"description" : ""
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<s-----\u003E\n  <syllabusId>...<\/syllabusId>\n  <content>...<\/content>\n  <sectionId>...<\/sectionId>\n<\/s-----\u003E\n"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/xml_ns0_syllabusDTO",
"description" : ""
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
        ]
      },
      "put" : {
        "tags" : [ "SyllabusController" ],
        "operationId" : "updateSyllabus",
        "requestBody" : {
          "required" : false,
          "content" : {
            "application/json" : {
              "examples" : {
                "example" : {
                  "value" : "{\n  \"content\" : \"...\",\n  \"sectionId\" : 12345,\n  \"syllabusId\" : 12345\n}"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/SyllabusDTO",
"description" : ""
              }
            }, 
            "application/xml" : {
              "examples" : {
                "example" : {
                  "value" : "<s-----\u003E\n  <syllabusId>...<\/syllabusId>\n  <content>...<\/content>\n  <sectionId>...<\/sectionId>\n<\/s-----\u003E\n"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/xml_ns0_syllabusDTO",
"description" : ""
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "{\n  \"content\" : \"...\",\n  \"sectionId\" : 12345,\n  \"syllabusId\" : 12345\n}"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/SyllabusDTO",
"description" : ""
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<s-----\u003E\n  <syllabusId>...<\/syllabusId>\n  <content>...<\/content>\n  <sectionId>...<\/sectionId>\n<\/s-----\u003E\n"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/xml_ns0_syllabusDTO",
"description" : ""
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
        ]
      }
    },
    "\/api/syllabuses/generate/{sectionId}" : {
      "post" : {
        "tags" : [ "SyllabusController" ],
        "operationId" : "generateSyllabus",
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "{\n  \"content\" : \"...\",\n  \"sectionId\" : 12345,\n  \"syllabusId\" : 12345\n}"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/SyllabusDTO",
"description" : ""
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<s-----\u003E\n  <syllabusId>...<\/syllabusId>\n  <content>...<\/content>\n  <sectionId>...<\/sectionId>\n<\/s-----\u003E\n"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/xml_ns0_syllabusDTO",
"description" : ""
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "sectionId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      }
    },
    "\/api/syllabuses/{syllabusId}" : {
      "delete" : {
        "tags" : [ "SyllabusController" ],
        "operationId" : "deleteSyllabus",
        "responses" : {
          "204" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "syllabusId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      },
      "get" : {
        "tags" : [ "SyllabusController" ],
        "operationId" : "getSyllabusById",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "{\n  \"content\" : \"...\",\n  \"sectionId\" : 12345,\n  \"syllabusId\" : 12345\n}"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/SyllabusDTO",
"description" : ""
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<s-----\u003E\n  <syllabusId>...<\/syllabusId>\n  <content>...<\/content>\n  <sectionId>...<\/sectionId>\n<\/s-----\u003E\n"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/xml_ns0_syllabusDTO",
"description" : ""
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "syllabusId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      }
    }
  }
}